Move hotplugged volumes and container disks
Move the hotplugged volumes and container disks in the mount manager
directory.
