+++ pkg/virt-handler/mount-manager/BUILD.bazel
@@ -8,9 +8,9 @@ go_library(
     deps = [
         "//pkg/container-disk:go_default_library",
         "//pkg/virt-config:go_default_library",
-        "//pkg/virt-handler/container-disk:go_default_library",
-        "//pkg/virt-handler/hotplug-disk:go_default_library",
         "//pkg/virt-handler/isolation:go_default_library",
+        "//pkg/virt-handler/mount-manager/container-disk:go_default_library",
+        "//pkg/virt-handler/mount-manager/hotplug-disk:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
     ],
