+++ pkg/virt-handler/mount-manager/hotplug-disk/BUILD.bazel
@@ -3,24 +3,28 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
+        "findmnt.go",
         "generated_mock_mount.go",
         "mount.go",
     ],
-    importpath = "kubevirt.io/kubevirt/pkg/virt-handler/container-disk",
+    importpath = "kubevirt.io/kubevirt/pkg/virt-handler/mount-manager/hotplug-disk",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/container-disk:go_default_library",
         "//pkg/ephemeral-disk-utils:go_default_library",
+        "//pkg/hotplug-disk:go_default_library",
         "//pkg/safepath:go_default_library",
+        "//pkg/storage/types:go_default_library",
         "//pkg/unsafepath:go_default_library",
-        "//pkg/util:go_default_library",
-        "//pkg/virt-config:go_default_library",
+        "//pkg/virt-handler/cgroup:go_default_library",
         "//pkg/virt-handler/isolation:go_default_library",
         "//pkg/virt-handler/virt-chroot:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/github.com/golang/mock/gomock:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
+        "//vendor/github.com/opencontainers/runc/libcontainer/configs:go_default_library",
+        "//vendor/github.com/opencontainers/runc/libcontainer/devices:go_default_library",
+        "//vendor/github.com/pkg/errors:go_default_library",
+        "//vendor/golang.org/x/sys/unix:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
     ],
 )
@@ -28,19 +32,30 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
-        "container_disk_suite_test.go",
+        "findmnt_test.go",
+        "hotplug-disk_suite_test.go",
         "mount_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
-        "//pkg/container-disk:go_default_library",
         "//pkg/ephemeral-disk-utils:go_default_library",
-        "//pkg/testutils:go_default_library",
+        "//pkg/hotplug-disk:go_default_library",
+        "//pkg/safepath:go_default_library",
+        "//pkg/unsafepath:go_default_library",
+        "//pkg/virt-handler/cgroup:go_default_library",
+        "//pkg/virt-handler/isolation:go_default_library",
+        "//pkg/virt-handler/mount-manager/utils:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/api:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
+        "//vendor/github.com/golang/mock/gomock:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
+        "//vendor/github.com/opencontainers/runc/libcontainer/configs:go_default_library",
+        "//vendor/github.com/opencontainers/runc/libcontainer/devices:go_default_library",
+        "//vendor/golang.org/x/sys/unix:go_default_library",
+        "//vendor/k8s.io/api/core/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
     ],
 )
