+++ pkg/virt-handler/mount-manager/hotplug-disk/generated_mock_mount.go
