+++ tools/analyzers/banncheck/banncheck.json
@@ -28,11 +28,11 @@
         },
         {
           "justification ": "XXX: mount entries are recoreded in a separate file with absolute files. No additional risk, but confusing from an audit perspective, must be changed at some point.",
-          "allowedPkg": "kubevirt.io/kubevirt/pkg/virt-handler/container-disk"
+          "allowedPkg": "kubevirt.io/kubevirt/pkg/virt-handler/mount-manager/container-disk"
         },
         {
           "justification ": "XXX: mount entries are recoreded in a separate file with absolute files. No additional risk, but confusing from an audit perspective, must be changed at some point.",
-          "allowedPkg": "kubevirt.io/kubevirt/pkg/virt-handler/hotplug-disk"
+          "allowedPkg": "kubevirt.io/kubevirt/pkg/virt-handler/mount-manager/hotplug-disk"
         },
         {
           "justification ": "for selinux labeling safepaths need to be deconstructed to hand them over to the virt-chroot binary",
